git-svn-id: svn://svn.code.sf.net/p/statechum/code/XMachineTool/trunk@684 4b8d0e13-f52a-0410-bc0e-8c3b9955e25a

+++ src/statechum/JUConstants.java
@@ -72,7 +72,7 @@ public enum JUConstants {
 		 * @param compat value to convert
 		 * @return resulting value
 		 */
-		public static PAIRCOMPATIBILITY compatibilityToJUConstants(int compat)
+		public static PAIRCOMPATIBILITY compatibilityToJUConstants(long compat)
 		{
 			PAIRCOMPATIBILITY result = null;
 			if (compat == PAIRCOMPATIBILITY.INCOMPATIBLE.getInteger()) result = PAIRCOMPATIBILITY.INCOMPATIBLE;
